Correct for Pandas Bug-fix in 16.2
Pandas changed how it loads in empty CSV files via read_csv. With 16.0
and before, read_csv just created an empty DataFrame when an empty CSV
was passed in. After 16.2, read_csv is unable to handle empty CSV files,
thus I used this try/except block, looking for a TypeError exception,
which is caused from the names= variable on the empty CSV. Without the
name= variable, a ValueError would occur on an empty CSV.